Added: New tasklistClass.js to separate class builder from index.js
